services/ingestion/requirements.txt
requests
psycopg2-binary
sqlalchemy
tenacity
python-dotenv
